+++ preload.js
@@ -2,8 +2,10 @@ const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('usage', {
   get: () => ipcRenderer.invoke('usage:get'),
+  context: () => ipcRenderer.invoke('context:get'),
   version: () => ipcRenderer.invoke('app:version'),
   onUpdate: (cb) => ipcRenderer.on('usage:update', (_e, data) => cb(data)),
+  onContext: (cb) => ipcRenderer.on('context:update', (_e, data) => cb(data)),
   onError: (cb) => ipcRenderer.on('usage:error', (_e, msg) => cb(msg)),
   refresh: () => ipcRenderer.send('widget:refresh'),
   close: () => ipcRenderer.send('widget:close'),